feat: add wmo.WMOProduct class
